Advancing the Blog 22 - Delete Permissions & Http Status Codes

+++ comments/views.py
@@ -1,6 +1,6 @@
 from django.contrib import messages
 from django.contrib.contenttypes.models import ContentType
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, Http404, HttpResponse
 from django.shortcuts import render, get_object_or_404
 
 from .forms import CommentForm
@@ -8,7 +8,19 @@ from .models import Comment
 
 
 def comment_delete(request, id):
-    obj = get_object_or_404(Comment, id=id)
+    # obj = get_object_or_404(Comment, id=id)
+    try:
+        obj = Comment.objects.get(id=id)
+    except:
+        raise Http404
+    if obj.user != request.user:
+        # messages.error(request, "You don't have permissions to do this.")
+        # raise Http404
+        response = HttpResponse("You don't have permissions to do this.")
+        response.status_code = 403
+        return response
+        # return render(request, 'confirm_delete.html', context, status_code=403)
+
     if request.method == "POST":
         parent_obj_url = obj.content_object.get_absolute_url()
         obj.delete()
@@ -21,7 +33,15 @@ def comment_delete(request, id):
 
 
 def comment_thread(request, id):
-    obj = get_object_or_404(Comment, id=id)
+    # obj = get_object_or_404(Comment, id=id)
+    try:
+        obj = Comment.objects.get(id=id)
+    except:
+        raise Http404
+
+    if not obj.is_parent:
+        obj = obj.parent
+
     initial_data = {
         'content_type': obj.content_type,
         'object_id': obj.object_id,
